requirements.txt
faiss-cpu>=1.8.0
numpy>=1.26.0
open-clip-torch>=2.24.0
pdfplumber>=0.11.0
pillow>=10.0.0
python-docx>=1.1.0
sentence-transformers>=2.7.0
streamlit>=1.35.0
torch>=2.2.0
tqdm>=4.66.0
